+++ aiohttp/test_utils.py
@@ -71,9 +71,6 @@ def run_server(loop, *, listen_addr=('127.0.0.1', 0),
             if properties.get('close', False):
                 return
 
-            if properties.get('noresponse', False):
-                yield from asyncio.sleep(99999)
-
             for hdr, val in message.headers.items():
                 if (hdr == 'EXPECT') and (val == '100-continue'):
                     self.transport.write(b'HTTP/1.0 100 Continue\r\n\r\n')