+++ aiohttp/web_exceptions.py
@@ -20,6 +20,7 @@ __all__ = (
     'HTTPNotModified',
     'HTTPUseProxy',
     'HTTPTemporaryRedirect',
+    'HTTPPermanentRedirect',
     'HTTPClientError',
     'HTTPBadRequest',
     'HTTPUnauthorized',
@@ -39,6 +40,11 @@ __all__ = (
     'HTTPUnsupportedMediaType',
     'HTTPRequestRangeNotSatisfiable',
     'HTTPExpectationFailed',
+    'HTTPMisdirectedRequest',
+    'HTTPUpgradeRequired',
+    'HTTPPreconditionRequired',
+    'HTTPTooManyRequests',
+    'HTTPRequestHeaderFieldsTooLarge',
     'HTTPServerError',
     'HTTPInternalServerError',
     'HTTPNotImplemented',
@@ -46,6 +52,9 @@ __all__ = (
     'HTTPServiceUnavailable',
     'HTTPGatewayTimeout',
     'HTTPVersionNotSupported',
+    'HTTPVariantAlsoNegotiates',
+    'HTTPNotExtended',
+    'HTTPNetworkAuthenticationRequired',
 )
 
 
@@ -163,6 +172,10 @@ class HTTPTemporaryRedirect(_HTTPMove):
     status_code = 307
 
 
+class HTTPPermanentRedirect(_HTTPMove):
+    status_code = 308
+
+
 ############################################################
 # 4xx client error
 ############################################################
@@ -253,6 +266,26 @@ class HTTPExpectationFailed(HTTPClientError):
     status_code = 417
 
 
+class HTTPMisdirectedRequest(HTTPClientError):
+    status_code = 421
+
+
+class HTTPUpgradeRequired(HTTPClientError):
+    status_code = 426
+
+
+class HTTPPreconditionRequired(HTTPClientError):
+    status_code = 428
+
+
+class HTTPTooManyRequests(HTTPClientError):
+    status_code = 429
+
+
+class HTTPRequestHeaderFieldsTooLarge(HTTPClientError):
+    status_code = 431
+
+
 ############################################################
 # 5xx Server Error
 ############################################################
@@ -291,3 +324,15 @@ class HTTPGatewayTimeout(HTTPServerError):
 
 class HTTPVersionNotSupported(HTTPServerError):
     status_code = 505
+
+
+class HTTPVariantAlsoNegotiates(HTTPServerError):
+    status_code = 506
+
+
+class HTTPNotExtended(HTTPServerError):
+    status_code = 510
+
+
+class HTTPNetworkAuthenticationRequired(HTTPServerError):
+    status_code = 511