+++ tests/test_signals.py
@@ -22,6 +22,8 @@ def debug_app(loop):
 
 def make_request(app, method, path, headers=CIMultiDict()):
     message = RawRequestMessage(method, path, HttpVersion11, headers,
+                                [(k.encode('utf-8'), v.encode('utf-8'))
+                                 for k, v in headers.items()],
                                 False, False)
     return request_from_message(message, app)
 
@@ -117,9 +119,15 @@ def test_copy_forbidden(app):
 
 
 def test_sort_forbidden(app):
-    l1 = lambda: None
-    l2 = lambda: None
-    l3 = lambda: None
+    def l1():
+        pass
+
+    def l2():
+        pass
+
+    def l3():
+        pass
+
     signal = Signal(app)
     signal.extend([l1, l2, l3])
     with pytest.raises(NotImplementedError):